🌈 Update(#30): organize some pages
